renamed `utils` module due to inconsistent behaviour in docs
