+++ networkcommons/eval/_metrics.py
@@ -41,7 +41,7 @@ import networkx as nx
 import decoupler as dc
 import numpy as np
 
-import networkcommons._utils as utils
+import networkcommons.utils as utils
 
 import random
