+++ networkcommons/utils.py
